Use background context for saving Meme to Core Data persistent storage
A parent managed object context uses a background thread, and
a child context uses the main thread.

The child context discards changes in case of an abandoned edit.

The MemeList singleton (used by the two navigation controller root
view controllers) keeps a reference to the child context; the detail
and editor view controllers are passed in a context to use.